+++ config/configmapprovider/default.go
@@ -14,9 +14,9 @@
 
 package configmapprovider // import "go.opentelemetry.io/collector/config/configmapprovider"
 
-// NewDefaultMapProvider returns the default MapProvider, and it creates configuration from a file
+// NewDefaultMapProvider returns the default Provider, and it creates configuration from a file
 // defined by the given configFile and overwrites fields using properties.
-func NewDefaultMapProvider(configFileName string, properties []string) MapProvider {
+func NewDefaultMapProvider(configFileName string, properties []string) Provider {
 	return NewExpandMapProvider(
 		NewMergeMapProvider(
 			NewFileMapProvider(configFileName),